+++ src/main/java/com/feed_the_beast/ftbquests/net/edit/MessageMoveChapterResponse.java
@@ -16,14 +16,14 @@ import net.minecraftforge.fml.relauncher.SideOnly;
  */
 public class MessageMoveChapterResponse extends MessageToClient
 {
-	private int id;
+	private short id;
 	private boolean up;
 
 	public MessageMoveChapterResponse()
 	{
 	}
 
-	public MessageMoveChapterResponse(int i, boolean u)
+	public MessageMoveChapterResponse(short i, boolean u)
 	{
 		id = i;
 		up = u;
@@ -38,14 +38,14 @@ public class MessageMoveChapterResponse extends MessageToClient
 	@Override
 	public void writeData(DataOut data)
 	{
-		data.writeInt(id);
+		data.writeShort(id);
 		data.writeBoolean(up);
 	}
 
 	@Override
 	public void readData(DataIn data)
 	{
-		id = data.readInt();
+		id = data.readShort();
 		up = data.readBoolean();
 	}
